html/parsing/html5lib/html-tree/comments01.dat

#data
FOO<!-- BAR -->BAZ
#errors
no doctype
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR  -->
|     "BAZ"

#data
FOO<!-- BAR --!>BAZ
#errors
no doctype
bad comment
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR  -->
|     "BAZ"

#data
FOO<!-- BAR --   >BAZ
#errors
no doctype
bad comment
EOF
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR --   >BAZ -->

#data
FOO<!-- BAR -- <QUX> -- MUX -->BAZ
#errors
no doctype
bad comment
bad comment
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR -- <QUX> -- MUX  -->
|     "BAZ"

#data
FOO<!-- BAR -- <QUX> -- MUX --!>BAZ
#errors
no doctype
bad comment
bad comment
bad comment
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR -- <QUX> -- MUX  -->
|     "BAZ"

#data
FOO<!-- BAR -- <QUX> -- MUX -- >BAZ
#errors
no doctype
bad
bad
bad
bad
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR -- <QUX> -- MUX -- >BAZ -->

#data
FOO<!---->BAZ
#errors
no doctype
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  -->
|     "BAZ"

#data
FOO<!--->BAZ
#errors
no doctype
bad
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  -->
|     "BAZ"

#data
FOO<!-->BAZ
#errors
no doctype
bad
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  -->
|     "BAZ"

#data
<?xml version="1.0">Hi
#errors
no doctype
bogus comment
#document
| <!-- ?xml version="1.0" -->
| <html>
|   <head>
|   <body>
|     "Hi"

#data
<?xml version="1.0">
#errors
no doctype
bogus comment
#document
| <!-- ?xml version="1.0" -->
| <html>
|   <head>
|   <body>

#data
<?xml version
#errors
no doctype
bogus comment
#document
| <!-- ?xml version -->
| <html>
|   <head>
|   <body>

#data
FOO<!----->BAZ
#errors
no doctype
bad comment
#document
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!-- - -->
|     "BAZ"
